+++ NEWS.md
@@ -1,6 +1,11 @@
+# exercism 0.1.0.9000
+
+* Included an RStudio addin that supports the main API functionality implemented 
+  in the package.
+
+* Update function names and documentation 
+
+
 # exercism 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
-
-
-